@@ -26,44 +26,54 @@ classifiers = [
 keywords = ["taskiq", "tasks", "distributed", "async"]
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8.1"
 typing-extensions = ">=3.10.0.0"
-pydantic = "^1.6.2"
+pydantic = ">=1.0,<=3.0"
 importlib-metadata = "*"
 pycron = "^3.0.0"
-taskiq_dependencies = "^1.2.3"
-anyio = "^3"
+taskiq_dependencies = ">=1.3.1,<2"
+anyio = ">=3"
+packaging = ">=19"
 # For prometheus metrics
 prometheus_client = { version = "^0", optional = true }
 # For ZMQBroker
-pyzmq = { version = "^23.2.0", optional = true }
+pyzmq = { version = "^23.2.0", optional = true, markers = "python_version < '3.12'"}
 # For speed
-uvloop = { version = ">=0.16.0,<1", optional = true }
+uvloop = { version = ">=0.16.0,<1", optional = true, markers = "sys_platform != 'win32'" }
 # For hot-reload.
 watchdog = { version = "^2.1.9", optional = true }
 gitignore-parser = { version = "^0", optional = true }
+pytz = "*"
+orjson = { version = "^3.9.9", optional = true }
+msgpack = { version = "^1.0.7", optional = true }
+cbor2 = { version = "^5.4.6", optional = true }
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1.2"
+ruff = "^0.0.291"
 black = { version = "^22.6.0", allow-prereleases = true }
-flake8 = "^4.0.1"
-isort = "^5.10.1"
 mypy = "^1"
 pre-commit = "^2.20.0"
-yesqa = "^1.3.0"
-autoflake = "^1.4"
-wemake-python-styleguide = "^0.16.1"
 coverage = "^6.4.2"
 pytest-cov = "^3.0.0"
 mock = "^4.0.3"
 pytest-xdist = { version = "^2.5.0", extras = ["psutil"] }
 types-mock = "^4.0.15"
+tox = "^4.6.4"
+freezegun = "^1.2.2"
+pytest-mock = "^3.11.1"
+tzlocal = "^5.0.1"
+types-tzlocal = "^5.0.1.1"
+types-pytz = "^2023.3.1.1"
 
 [tool.poetry.extras]
 zmq = ["pyzmq"]
 uv = ["uvloop"]
 metrics = ["prometheus_client"]
 reload = ["watchdog", "gitignore-parser"]
+orjson = ["orjson"]
+msgpack = ["msgpack"]
+cbor = ["cbor2"]
 
 [tool.poetry.scripts]
 taskiq = "taskiq.__main__:main"
