@@ -1,10 +1,8 @@
 import json
-import pickle  # noqa: S403
-from functools import partial
-from typing import Any, Callable, Dict, Generic, Optional, TypeVar
+import pickle
+from typing import Any, Dict, Generic, Optional, TypeVar
 
-from pydantic import validator
-from pydantic.generics import GenericModel
+from pydantic import BaseModel, ConfigDict, Field, field_serializer, field_validator
 from typing_extensions import Self
 
 from taskiq.serialization import exception_to_python, prepare_exception
@@ -12,18 +10,7 @@ from taskiq.serialization import exception_to_python, prepare_exception
 _ReturnType = TypeVar("_ReturnType")
 
 
-def _json_encoder(value: Any, default: Callable[[Any], Any]) -> Any:
-    if isinstance(value, BaseException):
-        return prepare_exception(value, json)
-
-    return default(value)
-
-
-def _json_dumps(value: Any, *, default: Callable[[Any], Any], **kwargs: Any) -> str:
-    return json.dumps(value, default=partial(_json_encoder, default=default), **kwargs)
-
-
-class TaskiqResult(GenericModel, Generic[_ReturnType]):
+class TaskiqResult(BaseModel, Generic[_ReturnType]):
     """Result of a remote task invocation."""
 
     is_err: bool
@@ -33,13 +20,24 @@ class TaskiqResult(GenericModel, Generic[_ReturnType]):
     log: Optional[str] = None
     return_value: _ReturnType
     execution_time: float
+    labels: Dict[str, str] = Field(default_factory=dict)
 
     error: Optional[BaseException] = None
 
-    class Config:
-        arbitrary_types_allowed = True
-        json_dumps = _json_dumps  # type: ignore
-        json_loads = json.loads
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
+    @field_serializer("error")
+    def serialize_error(self, value: BaseException) -> Any:
+        """
+        Serialize error field.
+
+        :returns: Any
+        :param value: exception to serialize.
+        """
+        if value:
+            return prepare_exception(value, json)
+
+        return None
 
     def raise_for_error(self) -> "Self":
         """Raise exception if `error`.
@@ -52,7 +50,7 @@ class TaskiqResult(GenericModel, Generic[_ReturnType]):
         return self
 
     def __getstate__(self) -> Dict[Any, Any]:
-        dict = super().__getstate__()  # noqa: WPS125
+        dict = super().__getstate__()
         vals: Dict[str, Any] = dict["__dict__"]
 
         if "error" in vals and vals["error"] is not None:
@@ -63,7 +61,7 @@ class TaskiqResult(GenericModel, Generic[_ReturnType]):
 
         return dict
 
-    @validator("error", pre=True)
+    @field_validator("error", mode="before")
     @classmethod
     def _validate_error(cls, value: Any) -> Optional[BaseException]:
         return exception_to_python(value)
