@@ -1,18 +1,28 @@
+import json
 import pickle
 
 import pytest
 
 from taskiq import TaskiqResult
+from taskiq.compat import model_dump_json
 
 
 def test_json_serialization() -> None:
-    task = TaskiqResult(is_err=False, return_value="some value", execution_time=0)
-    data: TaskiqResult[str] = TaskiqResult.parse_raw(task.json())
-    assert data.return_value == task.return_value
+    task: TaskiqResult[str] = TaskiqResult(
+        is_err=False,
+        return_value="some value",
+        execution_time=0,
+    )
+    data = json.loads(model_dump_json(task))
+    assert data["return_value"] == task.return_value
 
 
 def test_pickle_serialization() -> None:
-    task = TaskiqResult(is_err=False, return_value="some value", execution_time=0)
+    task: TaskiqResult[str] = TaskiqResult(
+        is_err=False,
+        return_value="some value",
+        execution_time=0,
+    )
     data: TaskiqResult[str] = pickle.loads(pickle.dumps(task))
     assert data.return_value == task.return_value
 
@@ -23,11 +33,18 @@ def test_json_error_serialization() -> None:
     except Exception as exc:
         error = exc
 
-    task = TaskiqResult(is_err=False, return_value=1, execution_time=0, error=error)
-    data: TaskiqResult[int] = TaskiqResult.parse_raw(task.json())
+    task: TaskiqResult[int] = TaskiqResult(
+        is_err=False,
+        return_value=1,
+        execution_time=0,
+        error=error,
+    )
+    data = json.loads(model_dump_json(task))
 
-    assert data.error.args == task.error.args  # type: ignore
-    assert type(data.error) == type(task.error)
+    assert len(data["error"]["exc_message"]) == 2
+    args = list(task.error.args)  # type: ignore
+    assert data["error"]["exc_message"][0] == args[0]  # type: ignore
+    assert data["error"]["exc_message"][1] == args[1]  # type: ignore
 
 
 def test_pickle_error_serialization() -> None:
@@ -36,16 +53,26 @@ def test_pickle_error_serialization() -> None:
     except Exception as exc:
         error = exc
 
-    task = TaskiqResult(is_err=False, return_value=1, execution_time=0, error=error)
+    task: TaskiqResult[int] = TaskiqResult(
+        is_err=False,
+        return_value=1,
+        execution_time=0,
+        error=error,
+    )
     data = pickle.loads(pickle.dumps(task))
 
     assert data.error.args == task.error.args  # type: ignore
-    assert type(data.error) == type(task.error)
+    assert type(data.error) == type(task.error)  # noqa: E721
 
 
 def test_result_raise_for_error_exc() -> None:
     error = ValueError("Error")
-    res = TaskiqResult(is_err=True, return_value=None, execution_time=0, error=error)
+    res: TaskiqResult[None] = TaskiqResult(
+        is_err=True,
+        return_value=None,
+        execution_time=0,
+        error=error,
+    )
     res = pickle.loads(pickle.dumps(res))
 
     with pytest.raises(ValueError) as exc:
@@ -55,7 +82,7 @@ def test_result_raise_for_error_exc() -> None:
 
 
 def test_result_raise_for_error_res() -> None:
-    res = TaskiqResult(
+    res: TaskiqResult[str] = TaskiqResult(
         is_err=False,
         return_value="Some value",
         execution_time=0,
