@@ -5,12 +5,13 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 from functools import wraps
 from logging import getLogger
-from typing import (  # noqa: WPS235
+from typing import (
     TYPE_CHECKING,
     Any,
     AsyncGenerator,
     Awaitable,
     Callable,
+    ClassVar,
     DefaultDict,
     Dict,
     List,
@@ -24,11 +25,14 @@ from uuid import uuid4
 from typing_extensions import ParamSpec, Self, TypeAlias
 
 from taskiq.abc.middleware import TaskiqMiddleware
+from taskiq.abc.serializer import TaskiqSerializer
+from taskiq.acks import AckableMessage
 from taskiq.decor import AsyncTaskiqDecoratedTask
 from taskiq.events import TaskiqEvents
-from taskiq.formatters.json_formatter import JSONFormatter
+from taskiq.formatters.proxy_formatter import ProxyFormatter
 from taskiq.message import BrokerMessage
 from taskiq.result_backends.dummy import DummyResultBackend
+from taskiq.serializers.json_serializer import JSONSerializer
 from taskiq.state import TaskiqState
 from taskiq.utils import maybe_awaitable, remove_suffix
 from taskiq.warnings import TaskiqDeprecationWarning
@@ -37,7 +41,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from taskiq.abc.formatter import TaskiqFormatter
     from taskiq.abc.result_backend import AsyncResultBackend
 
-_T = TypeVar("_T")  # noqa: WPS111
+_T = TypeVar("_T")
 _FuncParams = ParamSpec("_FuncParams")
 _ReturnType = TypeVar("_ReturnType")
 
@@ -67,8 +71,7 @@ class AsyncBroker(ABC):
     in async mode.
     """
 
-    available_tasks: Dict[str, AsyncTaskiqDecoratedTask[Any, Any]] = {}
-    is_worker_process: bool = False
+    global_task_registry: ClassVar[Dict[str, AsyncTaskiqDecoratedTask[Any, Any]]] = {}
 
     def __init__(
         self,
@@ -82,6 +85,7 @@ class AsyncBroker(ABC):
                 "Setting result backend with constructor is deprecated. "
                 "Please use `with_result_backend` instead.",
                 TaskiqDeprecationWarning,
+                stacklevel=2,
             )
         if task_id_generator is None:
             task_id_generator = default_id_generator
@@ -90,12 +94,15 @@ class AsyncBroker(ABC):
                 "Setting id generator with constructor is deprecated. "
                 "Please use `with_id_generator` instead.",
                 TaskiqDeprecationWarning,
+                stacklevel=2,
             )
         self.middlewares: "List[TaskiqMiddleware]" = []
         self.result_backend = result_backend
         self.decorator_class = AsyncTaskiqDecoratedTask
-        self.formatter: "TaskiqFormatter" = JSONFormatter()
+        self.serializer: TaskiqSerializer = JSONSerializer()
+        self.formatter: "TaskiqFormatter" = ProxyFormatter(self)
         self.id_generator = task_id_generator
+        self.local_task_registry: Dict[str, AsyncTaskiqDecoratedTask[Any, Any]] = {}
         # Every event has a list of handlers.
         # Every handler is a function which takes state as a first argument.
         # And handler can be either sync or async.
@@ -105,6 +112,46 @@ class AsyncBroker(ABC):
         ] = defaultdict(list)
         self.state = TaskiqState()
         self.custom_dependency_context: Dict[Any, Any] = {}
+        self.dependency_overrides: Dict[Any, Any] = {}
+        # True only if broker runs in worker process.
+        self.is_worker_process: bool = False
+        # True only if broker runs in scheduler process.
+        self.is_scheduler_process: bool = False
+
+    def find_task(self, task_name: str) -> Optional[AsyncTaskiqDecoratedTask[Any, Any]]:
+        """
+        Returns task by name.
+
+        This method should be used to get task by name.
+        Instead of accessing `available_tasks` or `local_available_tasks` directly.
+
+        It searches task by name in dict of tasks that
+        were registered for this broker directly.
+        If it fails, it checks global dict of all available tasks.
+
+        :param task_name: name of a task.
+        :returns: found task or None.
+        """
+        return self.local_task_registry.get(
+            task_name,
+        ) or self.global_task_registry.get(
+            task_name,
+        )
+
+    def get_all_tasks(self) -> Dict[str, AsyncTaskiqDecoratedTask[Any, Any]]:
+        """
+        Method to fetch all tasks available in broker.
+
+        This method returns all tasks, globally and locally
+        available in broker. With local tasks having higher priority.
+
+        So, if you have two tasks with the same name,
+        one registered in global registry and one registered
+        in local registry, then local task will be returned.
+
+        :return: dict of all tasks. Keys are task names, values are tasks.
+        """
+        return {**self.global_task_registry, **self.local_task_registry}
 
     def add_dependency_context(self, new_ctx: Dict[Any, Any]) -> None:
         """
@@ -145,6 +192,10 @@ class AsyncBroker(ABC):
         for handler in self.event_handlers[event]:
             await maybe_awaitable(handler(self.state))
 
+        for middleware in self.middlewares:
+            if middleware.__class__.startup != TaskiqMiddleware.startup:
+                await maybe_awaitable(middleware.startup())
+
         await self.result_backend.startup()
 
     async def shutdown(self) -> None:
@@ -161,6 +212,10 @@ class AsyncBroker(ABC):
         # Call all shutdown events.
         for handler in self.event_handlers[event]:
             await maybe_awaitable(handler(self.state))
+
+        for middleware in self.middlewares:
+            if middleware.__class__.shutdown != TaskiqMiddleware.shutdown:
+                await maybe_awaitable(middleware.shutdown)
 
         await self.result_backend.shutdown()
 
@@ -182,12 +237,18 @@ class AsyncBroker(ABC):
         """
 
     @abstractmethod
-    def listen(self) -> AsyncGenerator[bytes, None]:
+    def listen(self) -> AsyncGenerator[Union[bytes, AckableMessage], None]:
         """
         This function listens to new messages and yields them.
 
         This it the main point for workers.
         This function is used to get new tasks from the network.
+
+        If your broker support acknowledgement, then you
+        should wrap your message in AckableMessage dataclass.
+
+        If your messages was wrapped in AckableMessage dataclass,
+        taskiq will call ack when finish processing message.
 
         :yield: incoming messages.
         :return: nothing.
@@ -250,7 +311,7 @@ class AsyncBroker(ABC):
             def inner(
                 func: Callable[_FuncParams, _ReturnType],
             ) -> AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]:
-                nonlocal inner_task_name  # noqa: WPS420
+                nonlocal inner_task_name
                 if inner_task_name is None:
                     fmodule = func.__module__
                     if fmodule == "__main__":  # pragma: no cover
@@ -259,7 +320,10 @@ class AsyncBroker(ABC):
                                 os.path.sep,
                             ),
                         )
-                    inner_task_name = f"{fmodule}:{func.__name__}"  # noqa: WPS442
+                    fname = func.__name__
+                    if fname == "<lambda>":
+                        fname = f"lambda_{uuid4().hex}"
+                    inner_task_name = f"{fmodule}:{fname}"
                 wrapper = wraps(func)
 
                 decorated_task = wrapper(
@@ -271,7 +335,7 @@ class AsyncBroker(ABC):
                     ),
                 )
 
-                self.available_tasks[decorated_task.task_name] = decorated_task
+                self._register_task(decorated_task.task_name, decorated_task)
 
                 return decorated_task
 
@@ -288,6 +352,28 @@ class AsyncBroker(ABC):
             inner_task_name=task_name,
             inner_labels=labels or {},
         )
+
+    def register_task(
+        self,
+        func: Callable[_FuncParams, _ReturnType],
+        task_name: Optional[str] = None,
+        **labels: Any,
+    ) -> AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]:
+        """
+        API for registering tasks programmatically.
+
+        This function is basically the same as `task` decorator,
+        but it doesn't decorate function, it just registers it
+        and returns AsyncTaskiqDecoratedTask object, that can
+        be called later.
+
+        :param func: function to register.
+        :param task_name: custom name of a task, defaults to qualified function's name.
+        :param labels: some addition labels for task.
+
+        :returns: registered task.
+        """
+        return self.task(task_name=task_name, **labels)(func)
 
     def on_event(self, *events: TaskiqEvents) -> Callable[[EventHandler], EventHandler]:
         """
@@ -396,3 +482,42 @@ class AsyncBroker(ABC):
         """
         self.event_handlers[event].extend(handlers)
         return self
+
+    def with_serializer(
+        self,
+        serializer: TaskiqSerializer,
+    ) -> "Self":  # pragma: no cover
+        """
+        Set a new serializer and return an updated broker.
+
+        :param serializer: new serializer.
+        :return: self
+        """
+        self.serializer = serializer
+        return self
+
+    def with_formatter(self, formatter: "TaskiqFormatter") -> "Self":
+        """
+        Set new formatter and return an updated broker.
+
+        :param formatter: new formatter.
+        :return: self
+        """
+        self.formatter = formatter
+        return self
+
+    def _register_task(
+        self,
+        task_name: str,
+        task: AsyncTaskiqDecoratedTask[Any, Any],
+    ) -> None:
+        """
+        Mehtod is used to register tasks.
+
+        By default we register tasks in local task registry.
+        But this behaviour can be changed in subclasses.
+
+        :param task_name: Name of a task.
+        :param task: Decorated task.
+        """
+        self.local_task_registry[task_name] = task
