@@ -1,4 +1,4 @@
-from typing import AsyncGenerator, Optional, TypeVar
+from typing import Any, AsyncGenerator, Optional, TypeVar
 
 from typing_extensions import ParamSpec
 
@@ -71,5 +71,13 @@ class AsyncSharedBroker(AsyncBroker):
         """
         raise TaskiqError("Shared broker cannot listen")
 
+    def _register_task(
+        self,
+        task_name: str,
+        task: AsyncTaskiqDecoratedTask[Any, Any],
+    ) -> None:
+        self.global_task_registry[task_name] = task
+
 
 async_shared_broker = AsyncSharedBroker()
+shared_task = async_shared_broker.task
