@@ -7,12 +7,12 @@ from taskiq.abc.result_backend import AsyncResultBackend
 from taskiq.message import BrokerMessage
 
 try:
-    import zmq  # noqa: WPS433
-    from zmq.asyncio import Context  # noqa: WPS433
+    import zmq
+    from zmq.asyncio import Context, Socket
 except ImportError:
     zmq = None  # type: ignore
 
-_T = TypeVar("_T")  # noqa: WPS111
+_T = TypeVar("_T")
 
 logger = getLogger(__name__)
 
@@ -43,12 +43,23 @@ class ZeroMQBroker(AsyncBroker):
         self.context = Context()
         self.pub_host = zmq_pub_host
         self.sub_host = zmq_sub_host
+        self.socket: Socket
+
+    async def startup(self) -> None:
+        """
+        Startup for zmq broker.
+
+        This function creates actual connections to
+        sockets. if current process is worker,
+        it subscribes, otherwise it becomes publisher.
+        """
         if self.is_worker_process:
             self.socket = self.context.socket(zmq.SUB)
             self.socket.setsockopt(zmq.SUBSCRIBE, b"")
         else:
             self.socket = self.context.socket(zmq.PUB)
             self.socket.bind(self.pub_host)
+        await super().startup()
 
     async def kick(self, message: BrokerMessage) -> None:
         """
@@ -76,6 +87,6 @@ class ZeroMQBroker(AsyncBroker):
         :yields: incoming messages.
         """
         with self.socket.connect(self.sub_host) as sock:
-            while True:  # noqa: WPS457
+            while True:
                 data = await sock.recv_multipart()
                 yield b"".join(data)
