@@ -4,7 +4,6 @@ from pathlib import Path
 from tempfile import gettempdir
 from typing import Any, Optional
 
-from taskiq.abc.broker import AsyncBroker
 from taskiq.abc.middleware import TaskiqMiddleware
 from taskiq.message import TaskiqMessage
 from taskiq.result import TaskiqResult
@@ -19,7 +18,9 @@ class PrometheusMiddleware(TaskiqMiddleware):
     This middleware starts wsgi server with prometheus metrics.
     Also it updates metrics on events.
 
-    :param server_port: meme
+    :param server_port: The port to listen on.
+    :param server_addr: The address to listen on.
+    :paam metrics_path: The path to store metrics for multiproc env.
     """
 
     def __init__(
@@ -30,15 +31,6 @@ class PrometheusMiddleware(TaskiqMiddleware):
     ) -> None:
         super().__init__()
 
-        self.found_errors = None
-        self.received_tasks = None
-        self.success_tasks = None
-        self.saved_results = None
-        self.execution_time = None
-
-        if not AsyncBroker.is_worker_process:
-            return
-
         metrics_path = metrics_path or Path(gettempdir()) / "taskiq_worker"
 
         if not metrics_path.exists():
@@ -47,16 +39,12 @@ class PrometheusMiddleware(TaskiqMiddleware):
         logger.debug(f"Setting up multiproc dir to {metrics_path}")
 
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = str(metrics_path)
-        os.environ["prometheus_multiproc_dir"] = str(metrics_path)
+        os.environ["PROMETHEUS_MULTIPROC_DIR"] = str(metrics_path)
 
         logger.debug("Initializing metrics")
 
         try:
-            from prometheus_client import (  # noqa: WPS433
-                Counter,
-                Histogram,
-                start_http_server,
-            )
+            from prometheus_client import Counter, Histogram
         except ImportError as exc:
             raise ImportError(
                 "Cannot initialize metrics. Please install 'taskiq[metrics]'.",
@@ -84,13 +72,26 @@ class PrometheusMiddleware(TaskiqMiddleware):
         )
         self.execution_time = Histogram(
             "execution_time",
-            "Tome of function execution",
+            "Time of function execution",
             ["task_name"],
         )
-        try:
-            start_http_server(port=server_port, addr=server_addr)
-        except OSError as exc:
-            logger.debug("Cannot start prometheus server: %s", exc)
+        self.server_port = server_port
+        self.server_addr = server_addr
+
+    def startup(self) -> None:
+        """
+        Prometheus startup.
+
+        This function starts prometheus server.
+        It starts it only in case if it's a worker process.
+        """
+        from prometheus_client import start_http_server
+
+        if self.broker.is_worker_process:
+            try:
+                start_http_server(port=self.server_port, addr=self.server_addr)
+            except OSError as exc:
+                logger.debug("Cannot start prometheus server: %s", exc)
 
     def pre_execute(
         self,
@@ -105,8 +106,6 @@ class PrometheusMiddleware(TaskiqMiddleware):
         :param message: current message.
         :return: message
         """
-        if self.received_tasks is None:
-            return message
         self.received_tasks.labels(message.task_name).inc()
         return message
 
@@ -121,12 +120,6 @@ class PrometheusMiddleware(TaskiqMiddleware):
         :param message: received message.
         :param result: result of the execution.
         """
-        if (  # noqa: WPS337
-            self.success_tasks is None
-            or self.execution_time is None
-            or self.found_errors is None
-        ):
-            return
         if result.is_err:
             self.found_errors.labels(message.task_name).inc()
         else:
@@ -144,6 +137,4 @@ class PrometheusMiddleware(TaskiqMiddleware):
         :param message: received message.
         :param result: result of execution.
         """
-        if self.saved_results is None:
-            return
         self.saved_results.labels(message.task_name).inc()
