@@ -3,6 +3,7 @@ from logging import getLogger
 from typing import Any
 
 from taskiq.abc.middleware import TaskiqMiddleware
+from taskiq.exceptions import NoResultError
 from taskiq.message import TaskiqMessage
 from taskiq.result import TaskiqResult
 
@@ -15,8 +16,12 @@ class SimpleRetryMiddleware(TaskiqMiddleware):
     def __init__(
         self,
         default_retry_count: int = 3,
+        default_retry_label: bool = False,
+        no_result_on_retry: bool = True,
     ) -> None:
         self.default_retry_count = default_retry_count
+        self.default_retry_label = default_retry_label
+        self.no_result_on_retry = no_result_on_retry
 
     async def on_error(
         self,
@@ -37,24 +42,34 @@ class SimpleRetryMiddleware(TaskiqMiddleware):
         :param result: execution result.
         :param exception: found exception.
         """
+        # Valid exception
+        if isinstance(exception, NoResultError):
+            return
+
         retry_on_error = message.labels.get("retry_on_error")
+        if retry_on_error is None:
+            retry_on_error = "true" if self.default_retry_label else "false"
         # Check if retrying is enabled for the task.
-        if retry_on_error is None or retry_on_error.lower() != "true":
+        if retry_on_error.lower() != "true":
             return
         new_msg = deepcopy(message)
+
         # Getting number of previous retries.
         retries = int(new_msg.labels.get("_retries", 0)) + 1
         new_msg.labels["_retries"] = str(retries)
         max_retries = int(new_msg.labels.get("max_retries", self.default_retry_count))
+
         if retries < max_retries:
             logger.info(
                 "Task '%s' invocation failed. Retrying.",
                 message.task_name,
             )
-            new_msg.labels["_parent"] = message.task_id
-            new_msg.task_id = self.broker.id_generator()
             broker_message = self.broker.formatter.dumps(message=new_msg)
             await self.broker.kick(broker_message)
+
+            if self.no_result_on_retry:
+                result.error = NoResultError()
+
         else:
             logger.warning(
                 "Task '%s' invocation failed. Maximum retries count is reached.",
