@@ -1,6 +1,5 @@
 """Package for message receiver."""
+
 from taskiq.receiver.receiver import Receiver
 
-__all__ = [
-    "Receiver",
-]
+__all__ = ["Receiver"]
