@@ -1,8 +1,11 @@
+from logging import getLogger
 from typing import List
 
 from taskiq.abc.broker import AsyncBroker
 from taskiq.abc.schedule_source import ScheduleSource
-from taskiq.scheduler.scheduler import ScheduledTask
+from taskiq.scheduler.scheduled_task import ScheduledTask
+
+logger = getLogger(__name__)
 
 
 class LabelScheduleSource(ScheduleSource):
@@ -24,7 +27,7 @@ class LabelScheduleSource(ScheduleSource):
         :return: list of schedules.
         """
         schedules = []
-        for task_name, task in self.broker.available_tasks.items():
+        for task_name, task in self.broker.get_all_tasks().items():
             if task.broker != self.broker:
                 continue
             for schedule in task.labels.get("schedule", []):
@@ -40,6 +43,29 @@ class LabelScheduleSource(ScheduleSource):
                         kwargs=schedule.get("kwargs", {}),
                         cron=schedule.get("cron"),
                         time=schedule.get("time"),
+                        cron_offset=schedule.get("cron_offset"),
                     ),
                 )
         return schedules
+
+    def post_send(self, scheduled_task: ScheduledTask) -> None:
+        """
+        Remove `time` schedule from task's scheduler list.
+
+        Task just have sent and won't be sent by that trigger anymore. Other triggers in
+        scheduler list left unchanged.
+
+        :param scheduled_task: task that just have sent
+        """
+        if scheduled_task.cron or not scheduled_task.time:
+            return  # it's scheduled task with cron label, do not remove this trigger.
+
+        for task_name, task in self.broker.get_all_tasks().items():
+            if task.broker != self.broker or scheduled_task.task_name != task_name:
+                continue
+
+            schedule_list = task.labels.get("schedule", []).copy()
+            for idx, schedule in enumerate(schedule_list):
+                if schedule.get("time") == scheduled_task.time:
+                    task.labels.get("schedule", []).pop(idx)
+                    return
